nr-gnb-net-device: Fix ARFCN return type

+++ model/nr-gnb-net-device.h
@@ -66,7 +66,7 @@ class NrGnbNetDevice : public NrNetDevice
      */
     void SetCellId(uint16_t cellId);
 
-    uint16_t GetArfcn(uint8_t index) const;
+    uint32_t GetArfcn(uint8_t index) const;
 
     void SetRrc(Ptr<NrGnbRrc> rrc);
